refactor(runtime): extract synthesis-cache and archive from context-budget
Move the synthesis-cache, archived-tool-result archive, and runtime-event
history-search subsystems out of context-budget.ts into synthesis-cache.ts,
keeping behavior neutral:

- extract ~1500 lines (73 symbols: synthesis types/cache, archive read+write,
  stale-tool-result pruning, history search) to synthesis-cache.ts
- break the ContextBudgetPolicy reverse import by narrowing
  pruneStaleToolResultsBeforeCompact and collectStaleToolResultArchiveCandidates
  to accept the prune sub-policy + charsPerToken + minRecentTurns instead of
  the whole ContextBudgetPolicy
- export the 7 archive/synthesis helpers used by the still-inline history-
  compact engine, and re-export all public symbols from context-budget.ts for
  import compatibility

No reverse imports: synthesis-cache only imports the leaf
context-budget-helpers and @maka/core; context-budget and ai-sdk-backend
import forward.

Part of #1084.
